+++ components/HomePage/Header/HeaderList/HeaderList.module.scss
@@ -1,52 +1,28 @@
 @import '../../../../styles/utils/_imports.scss';
 
-$flyIn-animation-duration: 0.3s;
-@keyframes flyInEffect {
-  0% {
-    scale: 0;
-    transform: translateY(-10rem);
-  }
-  100% {
-    scale: 1;
-    transform: translateY(5.2rem);
-  }
-}
-
 .HeaderList {
   padding: 0 5rem;
   display: flex;
-  justify-content: space-between;
+  justify-content: center;
   align-items: center;
-  gap: 4rem;
-  order: 1;
-  @include respond(tablet) {
-    gap: 2rem;
-    flex: none;
-    order: 1;
+  gap: 5rem;
+  @include respond-max(mini-tablet) {
+    gap: 3rem;
   }
-  @include respond(mobile) {
+  @include respond-max(mobile) {
     display: none;
   }
+  flex: auto;
 }
 
 .HeaderListElement {
   font-family: $railway-font;
   font-style: normal;
   font-weight: 600;
-  @include respond(small-laptop) {
-    font-size: 2.5rem;
-    line-height: 4rem;
-  }
-  @include respond(tablet) {
-    font-size: 2rem;
-    line-height: 4rem;
-  }
-  @include respond(mobile) {
-    font-size: 1.5rem;
-    line-height: 3rem;
-  }
-
   font-size: 2rem;
+  @include respond-max(mini-tablet) {
+    font-size: 1.8rem;
+  }
   line-height: 4rem;
   letter-spacing: 0.05em;
   color: $dark-gray-font;
@@ -57,72 +33,5 @@ $flyIn-animation-duration: 0.3s;
   }
   &Active {
     color: $toggle-switch-color-dark;
-  }
-}
-
-.ViewList {
-  order: 1;
-  @include respond(medium-laptop) {
-    display: none;
-  }
-  @include respond(small-laptop) {
-    display: none;
-  }
-  @include respond(tablet) {
-    display: none;
-  }
-  @include respond(mobile) {
-    flex: none;
-    order: 2;
-    display: flex;
-  }
-  text-align: end;
-  display: flex;
-  justify-content: flex-end;
-  padding: 0 1rem;
-  &Icon {
-    font-size: 4rem;
-    color: $dark-gray-font;
-    cursor: pointer;
-  }
-}
-
-.RouterListDropDown {
-  position: absolute;
-  animation-duration: $flyIn-animation-duration;
-  animation-name: flyInEffect;
-  animation-timing-function: ease-in;
-
-  padding: 0.5rem 3.5rem;
-  box-shadow: -4px 4px 10px 7px rgba(0, 0, 0, 0.25);
-  border-radius: 0px 0px 6px 6px;
-  right: 0;
-  transform: translateY(5.2rem);
-  background: $light-bg;
-  display: flex;
-  flex-direction: column;
-  &_Dark {
-    background: $faded-dark-bg;
-  }
-  &_Elements {
-    font-family: $railway-font;
-    font-style: normal;
-    font-weight: 600;
-    font-size: 2.4rem;
-    line-height: 6rem;
-    letter-spacing: 0.05em;
-    color: $dark-gray-font;
-    cursor: pointer;
-    transition: 0.3s ease-in;
-    &:hover {
-      transform: scale(0.92);
-      color: $light-blue;
-    }
-    &_Active {
-      color: $toggle-switch-color-dark;
-    }
-    &_ActiveDark {
-      color: $light-blue;
-    }
   }
 }
